+++ CTS_Analytics/Models/Mnemonic/kuzModel.cs
@@ -37,6 +37,6 @@ namespace CTS_Analytics.Models.Mnemonic
         public Mine_sklad Sklad { get; set; }
         public RaspoznModel Raspozn { get; set; }
 
-        public override decimal ProdFact { get { return (decimal)SkipPos1.TotalTonnsPerTimeInterval + (decimal)SkipPos2.TotalTonnsPerTimeInterval; } }
+        public override int ProdFact { get { return (int)(SkipPos1.TotalTonnsPerTimeInterval + SkipPos2.TotalTonnsPerTimeInterval); } }
     }
 }
