+++ CTS_Analytics/Models/Mnemonic/sar3Model.cs
@@ -30,6 +30,6 @@
         public Mine_rockUtil RockUtil { get; set; }
         public Mine_Kotel Kotel { get; set; }
 
-        public override decimal ProdFact { get { return (decimal)Skip.TotalTonnsPerTimeInterval; } }
+        public override int ProdFact { get { return (int)Skip.TotalTonnsPerTimeInterval; } }
     }
 }
